feat(exp002): complete pilot with integrity-verified adapter; correct dtype/engine contracts at source
EXP-002 pilot (Kaggle vokaigharibo/gharibo-exp-002-pilot version 6) reached COMPLETE.
Recovered artifact passes 29/29 integrity checks; adapter sha256 8d80d7ca...733e.

Contract corrections (all found by real run failures, all fixed at source):
- dtype: package builder hardcoded fp16 while the recipe/notebook declared float32;
  builder now reads the governed recipe. Shared package/worker types widened; the
  validator admits float32 and still rejects bf16 (Turing has no bf16 units).
- sequence_length 3072 admitted as a measured, allow-listed governed value.
- engine freeze: dependencies were empty, so `uv pip install` ran with no package
  argument and exited 2. The governed 12-dependency freeze set (DEC-0025) is restored.
- Unsloth's loader accepts only None/fp16/bf16, so the governed dtype is declared and
  recorded rather than forced through that argument.
- TRL strips `attention_mask`, so the assistant-only collator now tolerates its absence.
- loader revision pinned; governed chat template committed and hash-asserted.

New fail-closed gates: packageDtypeAgreesAcrossArtifacts (parses the manifest the
notebook actually receives), engineDependencyListNonEmpty/Complete,
loaderDtypeArgIsEngineCompatible, collatorToleratesMissingAttentionMask.

Governance: DEC-0049..DEC-0055; master state 1.28.0 -> 1.34.0; gate layers added
strip-and-delegate. EXP-001 failure history and the pilot's non-promotable origin
are preserved, not rewritten. Sealed qualification split untouched.

Validation: master:validate, docs:validate, verify:m2, eval:verify (76/76),
eval:audit:history (6/6), typecheck, 279/279 tests, git diff --check.

+++ apps/web/lib/workers/kaggle/index.ts
@@ -32,7 +32,11 @@ export {
 /** Kaggle free-tier capabilities (m2-stack-facts.md §1). */
 export const KAGGLE_CAPABILITIES: WorkerCapabilities = {
   gpuClass: "NVIDIA T4 (16 GB, sm_75)",
-  dtype: "fp16",
+  // gpt-oss on Turing: Unsloth refuses fp16 and switches to float32
+  // ("Using float16 precision for gpt_oss won't work! Using float32"), so the
+  // declared capability is the dtype the engine actually honours. Declaring
+  // fp16 here was the class of contradiction that failed pilot Version 1.
+  dtype: "float32",
   supportsResume: true,
   supportsSecrets: true,
   persistentPath: "/kaggle/working",
